feat: Change `Category.DisplayOrder` property from integer to decimal for finer-grained ordering.

+++ ProductCatalogSystem.Server/Features/Categories/Create/Models.cs
@@ -12,5 +12,5 @@ public class CreateCategoryRequest
 
     public CategoryStatus Status { get; set; } = CategoryStatus.Active;
 
-    public int DisplayOrder { get; set; }
+    public decimal DisplayOrder { get; set; }
 }
